RF: Resolving circular import by moving colour functions into psychopy.colors
