+++ psychopy/colors.py
@@ -6,7 +6,7 @@ from __future__ import absolute_import, print_function
 from past.builtins import basestring
 from psychopy.visual import basevisual as vis
 from past.builtins import basestring
-from psychopy.tools.colorspacetools import dkl2rgb, lms2rgb, hsv2rgb  # pylint: disable=W0611
+from psychopy.tools.coordinatetools import sph2cart
 from psychopy import logging
 import re
 import numpy
@@ -900,3 +900,134 @@ def isValidColor(color):
         return True
     else:
         return False
+
+def unpackColors(colors):  # used internally, not exported by __all__
+    """Reshape an array of color values to Nx3 format.
+
+    Many color conversion routines operate on color data in Nx3 format, where
+    rows are color space coordinates. 1x3 and NxNx3 input are converted to Nx3
+    format. The original shape and dimensions are also returned, allowing the
+    color values to be returned to their original format using 'reshape'.
+
+    Parameters
+    ----------
+    colors : ndarray, list or tuple of floats
+        Nx3 or NxNx3 array of colors, last dim must be size == 3 specifying each
+        color coordinate.
+
+    Returns
+    -------
+    tuple
+        Nx3 ndarray of converted colors, original shape, original dims.
+
+    """
+    # handle the various data types and shapes we might get as input
+    colors = numpy.asarray(colors, dtype=float)
+
+    orig_shape = colors.shape
+    orig_dim = colors.ndim
+    if orig_dim == 1 and orig_shape[0] == 3:
+        colors = numpy.array(colors, ndmin=2)
+    elif orig_dim == 2 and orig_shape[1] == 3:
+        pass  # NOP, already in correct format
+    elif orig_dim == 3 and orig_shape[2] == 3:
+        colors = numpy.reshape(colors, (-1, 3))
+    else:
+        raise ValueError(
+            "Invalid input dimensions or shape for input colors.")
+
+    return colors, orig_shape, orig_dim
+
+def dkl2rgb(dkl, conversionMatrix=None):
+    """Convert from DKL color space (Derrington, Krauskopf & Lennie) to RGB.
+
+    Requires a conversion matrix, which will be generated from generic
+    Sony Trinitron phosphors if not supplied (note that this will not be
+    an accurate representation of the color space unless you supply a
+    conversion matrix).
+
+    usage::
+
+        rgb(Nx3) = dkl2rgb(dkl_Nx3(el,az,radius), conversionMatrix)
+        rgb(NxNx3) = dkl2rgb(dkl_NxNx3(el,az,radius), conversionMatrix)
+
+    """
+    if conversionMatrix is None:
+        conversionMatrix = numpy.asarray([
+            # (note that dkl has to be in cartesian coords first!)
+            # LUMIN    %L-M    %L+M-S
+            [1.0000, 1.0000, -0.1462],  # R
+            [1.0000, -0.3900, 0.2094],  # G
+            [1.0000, 0.0180, -1.0000]])  # B
+        logging.warning('This monitor has not been color-calibrated. '
+                        'Using default DKL conversion matrix.')
+
+    if len(dkl.shape) == 3:
+        dkl_NxNx3 = dkl
+        # convert a 2D (image) of Spherical DKL colours to RGB space
+        origShape = dkl_NxNx3.shape  # remember for later
+        NxN = origShape[0] * origShape[1]  # find nPixels
+        dkl = numpy.reshape(dkl_NxNx3, [NxN, 3])  # make Nx3
+        rgb = dkl2rgb(dkl, conversionMatrix)  # convert
+        return numpy.reshape(rgb, origShape)  # reshape and return
+
+    else:
+        dkl_Nx3 = dkl
+        # its easier to use in the other orientation!
+        dkl_3xN = numpy.transpose(dkl_Nx3)
+        if numpy.size(dkl_3xN) == 3:
+            RG, BY, LUM = sph2cart(dkl_3xN[0],
+                                   dkl_3xN[1],
+                                   dkl_3xN[2])
+        else:
+            RG, BY, LUM = sph2cart(dkl_3xN[0, :],
+                                   dkl_3xN[1, :],
+                                   dkl_3xN[2, :])
+        dkl_cartesian = numpy.asarray([LUM, RG, BY])
+        rgb = numpy.dot(conversionMatrix, dkl_cartesian)
+
+        # return in the shape we received it:
+        return numpy.transpose(rgb)
+
+def lms2rgb(lms_Nx3, conversionMatrix=None):
+    """Convert from cone space (Long, Medium, Short) to RGB.
+
+    Requires a conversion matrix, which will be generated from generic
+    Sony Trinitron phosphors if not supplied (note that you will not get
+    an accurate representation of the color space unless you supply a
+    conversion matrix)
+
+    usage::
+
+        rgb_Nx3 = lms2rgb(dkl_Nx3(el,az,radius), conversionMatrix)
+
+    """
+
+    col = Color(tuple(lms_Nx3), 'lms')
+    if len(lms_Nx3) == 3:
+        return unpackColors(col.rgb)
+    elif len(lms_Nx3) == 4:
+        return unpackColors(col.rgba)
+
+def hsv2rgb(hsv_Nx3):
+    """Convert from HSV color space to RGB gun values.
+
+    usage::
+
+        rgb_Nx3 = hsv2rgb(hsv_Nx3)
+
+    Note that in some uses of HSV space the Hue component is given in
+    radians or cycles (range 0:1]). In this version H is given in
+    degrees (0:360).
+
+    Also note that the RGB output ranges -1:1, in keeping with other
+    PsychoPy functions.
+    """
+    # based on method in
+    # http://en.wikipedia.org/wiki/HSL_and_HSV#Converting_to_RGB
+
+    col = Color(tuple(hsv_Nx3), 'hsv')
+    if len(hsv_Nx3) == 3:
+        return unpackColors(col.rgb)
+    elif len(hsv_Nx3) == 4:
+        return unpackColors(col.rgba)
